+++ CHANGELOG.md
@@ -2,8 +2,16 @@
 
 ## Unreleased changes
 
+- Fix a bug in Ctrl-C signal handling where a node would fail to stop if
+  interrupted early on in the startup if out-of-band catchup was enabled.
 - `database-exporter` now produces a collection of export files, instead of a single file. The new
   `--chunksize` option specifies the size of export files in blocks.
+- The `--download-blocks-from` option now takes the URL to the catchup _index file_, permitting to
+  only download and import catchup files containing blocks not already present in the database.
+
+## 4.3.1
+
+- Improvements to start-up time that fix regressions introduced by the account caching.
 
 ## 4.3.0